成功在qemu9上利用自制sbi启动简易内核

+++ sbi-qemu/src/csr.rs
@@ -1,13 +1,13 @@
 use core::arch::asm;
 use paste::paste;
 // 常量定义
-pub const PRV_U: usize = 0;
-pub const PRV_S: usize = 1;
-pub const PRV_M: usize = 3;
+pub const PRV_U: u32 = 0;
+pub const PRV_S: u32 = 1;
+pub const PRV_M: u32 = 3;
 
-pub const MSTATUS_MPP_SHIFT: usize = 11;
-pub const MSTATUS_MPP: usize = 3 << MSTATUS_MPP_SHIFT;
-pub const MSTATUS_MPIE: usize = 0x00000080;
+pub const MSTATUS_MPP_SHIFT: u32 = 11;
+pub const MSTATUS_MPP: u32 = 3 << MSTATUS_MPP_SHIFT;
+pub const MSTATUS_MPIE: u32 = 0x00000080;
 /*
 例子：
 let mstatus = read_crs(mstatus);
@@ -21,7 +21,7 @@ write_csr(mstatus, new_mstatus);
 
 ///在 val 中插入 fieldval，插入的位置和大小由 which 决定。
 #[inline]
-pub fn insert_field(val: usize, which: usize, fieldval: usize) -> usize {
+pub fn insert_field(val: u32, which: u32, fieldval: u32) -> u32 {
     (val & !which) | (fieldval << which.trailing_zeros())
 }
 ///不能在运行时动态获取 CSR 名称,使用宏可以使csr名称在编译时确定
@@ -29,7 +29,7 @@ pub fn insert_field(val: usize, which: usize, fieldval: usize) -> usize {
 #[macro_export]
 macro_rules! read_csr {
     ($csr:expr) => {{
-        let value: usize;
+        let value: u32;
         unsafe {
             asm!(concat!("csrr {}, ", $csr), out(reg) value);
         }
@@ -52,12 +52,12 @@ macro_rules! define_csr_functions {
     ($csrname:ident) => {
         paste::paste! {
             #[inline]
-            pub fn [<read_ $csrname>]() -> usize {
+            pub fn [<read_ $csrname>]() -> u32 {
                 read_csr!(stringify!($csrname))
             }
 
             #[inline]
-            pub fn [<write_ $csrname>](val: usize) {
+            pub fn [<write_ $csrname>](val: u32) {
                 write_csr!(stringify!($csrname), val)
             }
         }
